refactor: move public assets to /assets folder to bypass auth middleware Adding auth required for all pages via middleware.ts
