+++ apps/endatix-hub/components/layout-ui/navigation/main-nav.tsx
@@ -8,7 +8,7 @@ import Image from "next/image";
 import NavLink from "./nav-link";
 import { sitemap } from "@/lib/constants";
 import { SitemapService } from "@/services/sitemap-service";
-import EndatixLogoSvg from "@/public/icons/icon.svg";
+import EndatixLogoSvg from "@/public/assets/icons/icon.svg";
 
 const MainNav = () => {
   const logo = SitemapService.getLogo();
